+++ api/src/main/java/com/toutiao/web/api/chance/response/advertisement/SellHouseResponse.java
@@ -16,12 +16,12 @@ public class SellHouseResponse {
     /**
      * 用户浏览的起始位置
      */
-    private Integer startBit;
+    private Long startBit;
 
     /**
      * 查询位置
      */
-    private Integer queryBit;
+    private Long queryBit;
 
     /**
      * 是否补充
